Added middleware; added virtual property
I added middleware to generate a JWT token which will be called when some user routes are called. Also, i added a toJSON method, which is run everytime res.json() is called to send json data to the client. The method removes the password and tokens array from the user instance before it's sent to the client

I also added a virtual property, Products, to the user schema.